now the project runs well with windows and visual studio

+++ src/renderer/swap_chain.h
@@ -24,7 +24,7 @@ public:
   SwapChain(const SwapChain &) = delete;
   SwapChain operator=(const SwapChain &) = delete;
 
-  VkFramebuffer getFrameBuffer(int iIndex) {
+  VkFramebuffer getFrameBuffer(size_t iIndex) {
     return mSwapChainFramebuffers[iIndex];
   }
   VkRenderPass getRenderPass() { return mRenderPass; }
